remove unused variables
